make voice golint compliant
