+++ .travis.yml
@@ -17,6 +17,7 @@ script:
   - $HOME/gopath/bin/golint sms/**/*.go
   - $HOME/gopath/bin/golint transcription/**/*.go
   - $HOME/gopath/bin/golint usage/**/*.go
+  - $HOME/gopath/bin/golint voice/**/*.go
   - $HOME/gopath/bin/golint */.go
 after_script:
   - FIXED=$(go fmt ./... | wc -l); if [ $FIXED -gt 0 ]; then echo "gofmt - $FIXED file(s) not formatted correctly, please run gofmt to fix this." && exit 1; fi
