+++ voice/calls.go
@@ -5,7 +5,6 @@ import (
 	"fmt"
 	"github.com/natebrennand/twiliogo/common"
 	"io"
-	"net/http"
 	"net/url"
 	"regexp"
 	"strconv"
@@ -13,28 +12,29 @@ import (
 	"time"
 )
 
+var (
+	validateCallSid      = regexp.MustCompile(`^CA[0-9a-z]{32}$`).MatchString
+	validateRecordingSid = regexp.MustCompile(`^RE[0-9a-z]{32}$`).MatchString
+)
+
+const (
+	postURL   = "https://api.twilio.com/2010-04-01/Accounts/%s/Calls.json"
+	updateURL = "https://api.twilio.com/2010-04-01/Accounts/%s/Calls/%s.json"
+	getURL    = "https://api.twilio.com/2010-04-01/Accounts/%s/Calls/%s.json"
+	listURL   = "https://api.twilio.com/2010-04-01/Accounts/%s/Calls.json"
+)
+
+// Account wraps the common Account struct to embed the AccountSid & Token.
 type Account struct {
-	AccountSid string
-	Token      string
-	Client     http.Client
+	common.Account
 }
 
-func (act Account) GetSid() string {
-	return act.AccountSid
-}
-func (act Account) GetToken() string {
-	return act.Token
-}
-func (act Account) GetClient() http.Client {
-	return act.Client
-}
-
-// Represents the data used in creating an outbound voice message.
+// Call represents the data used in creating an outbound voice message.
 // "From" & "To" are required attributes.
 // Either a ApplicationSid or a URL must also be provided.
 // Visit https://www.twilio.com/docs/api/rest/making-calls#post-parameters for more details and
 // explanation of other optional parameters.
-type Post struct {
+type Call struct {
 	From                 string
 	To                   string
 	Body                 string
@@ -50,7 +50,8 @@ type Post struct {
 	Record               *bool
 }
 
-func (p Post) GetReader() io.Reader {
+// GetReader implements the common.twilioPost interface
+func (p Call) GetReader() io.Reader {
 	vals := url.Values{}
 	vals.Set("To", p.To)
 	vals.Set("From", p.From)
@@ -88,10 +89,10 @@ func (p Post) GetReader() io.Reader {
 	return strings.NewReader(vals.Encode())
 }
 
-// Validates a Voice Post.
-func (p Post) Validate() error {
+// Validate implements the common.twilioPost interface
+func (p Call) Validate() error {
 	if p.From == "" || p.To == "" {
-		return errors.New("Both \"From\" and \"To\" must be set in Post.")
+		return errors.New("Both \"From\" and \"To\" must be set in Call.")
 	}
 	if p.ApplicationSid == "" && p.URL == "" {
 		return errors.New("Either \"ApplicationSid\" or \"URL\" must be set.")
@@ -99,21 +100,36 @@ func (p Post) Validate() error {
 	if p.SendDigits != "" {
 		match, err := regexp.MatchString(`^[0-9#\*w]+$`, p.SendDigits)
 		if match != true || err != nil {
-			return errors.New("Post's SendDigits can only contain digits, #, * or w")
+			return errors.New("Call's SendDigits can only contain digits, #, * or w")
 		}
 	}
 	return nil
 }
 
-// Sends a post request to Twilio to create a call.
-func (act Account) Call(p Post) (Call, error) {
-	var r Call
+// Call creates a new call with Twilio.
+func (act Account) Call(p Call) (Resource, error) {
+	var r Resource
 	err := common.SendPostRequest(fmt.Sprintf(postURL, act.AccountSid), p, act, &r)
 	return r, err
 }
 
-func (act Account) Get(sid string) (Call, error) {
-	var m Call
+// Resource represents a call record.
+type Resource struct {
+	common.ResponseCore
+	Price          common.JSONFloat `json:"price"`
+	ParentCallSid  string
+	PhoneNumberSid string
+	StartTime      common.JSONTime `json:"start_time"`
+	EndTime        common.JSONTime `json:"end_time"`
+	Duration       int64           `json:"duration,string"`
+	AnsweredBy     string          `json:"answered_by"`
+	ForwardedFrom  string          `json:"fowarded_from"`
+	CallerName     string          `json:"caller_name"`
+}
+
+// Get returns a call resource record.
+func (act Account) Get(sid string) (Resource, error) {
+	var m Resource
 	if !validateCallSid(sid) {
 		return m, errors.New("Invalid sid")
 	}
@@ -121,7 +137,74 @@ func (act Account) Get(sid string) (Call, error) {
 	return m, err
 }
 
-// Used to filter call logs results
+// Update is used to modify a call with Twiml.
+type Update struct {
+	URL                  string `json:"url"`
+	Method               string `json:"method"`
+	Status               string `json:"status"`
+	FallbackURL          string `json:"fallback_url"`
+	FallbackMethod       string `json:"fallback_method"`
+	StatusCallback       string `json:"status_callback"`
+	StatusCallbackMethod string `json:"status_callback_method"`
+}
+
+// GetReader implements the common.twilioPost interface
+func (p Update) GetReader() io.Reader {
+	vals := url.Values{}
+	if p.URL != "" {
+		vals.Set("URL", p.URL)
+	}
+	if p.Status != "" {
+		vals.Set("Status", p.Status)
+	}
+	if p.StatusCallback != "" {
+		vals.Set("StatusCallback", p.StatusCallback)
+	}
+	if p.StatusCallbackMethod != "" {
+		vals.Set("StatusCallbackMethod", p.StatusCallbackMethod)
+	}
+	if p.Method != "" {
+		vals.Set("Method", p.Method)
+	}
+	if p.FallbackURL != "" {
+		vals.Set("FallbackURL", p.FallbackURL)
+	}
+	if p.FallbackMethod != "" {
+		vals.Set("FallbackMethod", p.FallbackMethod)
+	}
+
+	return strings.NewReader(vals.Encode())
+}
+
+// Validate implements the common.twilioPost interface
+func (p Update) Validate() error {
+	if p.URL == "" && p.Method == "" && p.Status == "" {
+		return errors.New("URL or Status or Method must all be set")
+	}
+	return nil
+}
+
+// Internal function for sending the post request to twilio.
+func (act Account) postUpdate(dest string, msg Update, resp *Resource) error {
+	// send post request to twilio
+	return common.SendPostRequest(dest, msg, act, resp)
+}
+
+// Update sends an update to a Twilio call.
+func (act Account) Update(p Update, sid string) (Resource, error) {
+	var r Resource
+	err := common.SendPostRequest(fmt.Sprintf(updateURL, act.AccountSid, sid), p, act, &r)
+	return r, err
+}
+
+// CallList represents a list of call records.
+type CallList struct {
+	common.ListResponseCore
+	Calls *[]Resource `json:"calls"`
+	act   *Account
+}
+
+// ListFilter is used to filter call logs results
 type ListFilter struct {
 	To            string
 	From          string
@@ -155,8 +238,19 @@ func (f ListFilter) getQueryString() string {
 	return encoded
 }
 
+// List returns a list of Twilio call records
 func (act Account) List(f ListFilter) (CallList, error) {
 	var callList CallList
 	err := common.SendGetRequest(fmt.Sprintf(listURL, act.AccountSid)+f.getQueryString(), act, &callList)
+	callList.act = &act
 	return callList, err
+}
+
+// Next sets the MessageList to the next page of the list resource, returns an error in the
+// case that there are no more pages left.
+func (cl *CallList) Next() error {
+	if cl.Page == cl.NumPages-1 {
+		return errors.New("No more new pages")
+	}
+	return common.SendGetRequest(cl.NextPageURI, *cl.act, cl)
 }
